admin - securing views

+++ books/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
@@ -75,8 +76,12 @@ def book_detail(request, book_id):
 
     return render(request, 'books/book_detail.html', context)
 
-
+@login_required
 def add_book(request):
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store admin can do that')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = BookForm(request.POST, request.FILES)
         if form.is_valid():
@@ -94,8 +99,12 @@ def add_book(request):
 
     return render(request, template, context)
 
-
+@login_required
 def edit_book (request, book_id):
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store admin can do that')
+        return redirect(reverse('home'))
+
     book = get_object_or_404(Books, pk=book_id)
     if request.method == 'POST':
         form = BookForm(request.POST, request.FILES, instance=book)
@@ -117,8 +126,12 @@ def edit_book (request, book_id):
 
     return render(request, template, context)
 
-
+@login_required
 def delete_book(request, book_id):
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store admin can do that')
+        return redirect(reverse('home'))
+
     book = get_object_or_404(Books, pk=book_id)
     book.delete()
     messages.success(request, 'Book deleted')
